Fix: Can signup now

+++ create_tables.sql
@@ -1,14 +1,14 @@
 ﻿CREATE SCHEMA cs421g22;
 
 CREATE TABLE cs421g22.artists (
-    artist_id   BIGINT,
+    artist_id   INT,
     name        text NOT NULL,
     website     text,
     CONSTRAINT  pk_artists PRIMARY KEY ( artist_id )
 );
 
 CREATE TABLE cs421g22.collections (
-    collection_id   BIGINT,
+    collection_id   INT,
     name            text NOT NULL,
     genre           text ,
     creation_date   date DEFAULT current_date NOT NULL,
@@ -18,22 +18,22 @@ CREATE TABLE cs421g22.collections (
 COMMENT ON TABLE cs421g22.collections IS 'A generic collection of songs.';
 
 CREATE TABLE cs421g22.links (
-    link_id     BIGINT,
+    link_id     INT,
     url         text NOT NULL,
     source      varchar(100) NOT NULL,
     CONSTRAINT  pk_links PRIMARY KEY ( link_id )
 );
 
 CREATE TABLE cs421g22.songs (
-    song_id     BIGINT,
+    song_id     INT,
     name        text ,
-    length      BIGINT ,
+    length      INT ,
     CONSTRAINT  pk_songs PRIMARY KEY ( song_id )
 );
 
 
 CREATE TABLE cs421g22.users (
-    user_id     BIGINT,
+    user_id     INT,
     username    varchar(100) NOT NULL,
     password    text NOT NULL,
     super_user  bool DEFAULT 'false' NOT NULL,
@@ -41,8 +41,8 @@ CREATE TABLE cs421g22.users (
 );
 
 CREATE TABLE cs421g22.playlists (
-    playlist_id BIGINT,
-    likes_count BIGINT DEFAULT 0 NOT NULL,
+    playlist_id INT,
+    likes_count INT DEFAULT 0 NOT NULL,
     CONSTRAINT  pk_playlists PRIMARY KEY ( playlist_id ),
     CONSTRAINT  fk_playlists FOREIGN KEY ( playlist_id )
     REFERENCES  cs421g22.collections( collection_id )
@@ -50,7 +50,7 @@ CREATE TABLE cs421g22.playlists (
 );
 
 CREATE TABLE cs421g22.albums (
-    album_id    BIGINT NOT NULL,
+    album_id    INT NOT NULL,
     year        smallint ,
     song_count  smallint NOT NULL,
     metadata    text ,
@@ -61,8 +61,8 @@ CREATE TABLE cs421g22.albums (
 );
 
 CREATE TABLE cs421g22.users_subscriptions (
-    subscriber  BIGINT NOT NULL,
-    subscribee  BIGINT NOT NULL,
+    subscriber  INT NOT NULL,
+    subscribee  INT NOT NULL,
     since       timestamp DEFAULT current_timestamp ,
     CONSTRAINT  idx_users_subscriptions_0 UNIQUE ( subscriber, subscribee ) ,
     CONSTRAINT  fk_users_subscriptions FOREIGN KEY ( subscriber )
@@ -73,8 +73,8 @@ CREATE TABLE cs421g22.users_subscriptions (
 );
 
 CREATE TABLE cs421g22.song_has_link (
-    song_id     BIGINT NOT NULL,
-    link_id     BIGINT NOT NULL,
+    song_id     INT NOT NULL,
+    link_id     INT NOT NULL,
     CONSTRAINT  idx_song_has_link_1 UNIQUE ( song_id, link_id ) ,
     CONSTRAINT  fk_song_has_link FOREIGN KEY ( song_id )
     REFERENCES  cs421g22.songs( song_id )
@@ -85,8 +85,8 @@ CREATE TABLE cs421g22.song_has_link (
 );
 
 CREATE TABLE cs421g22.song_by_artist (
-    song_id     BIGINT NOT NULL,
-    artist_id   BIGINT NOT NULL,
+    song_id     INT NOT NULL,
+    artist_id   INT NOT NULL,
     CONSTRAINT  idx_song_by_artist UNIQUE ( song_id, artist_id ) ,
     CONSTRAINT  fk_album_by_artist_0 FOREIGN KEY ( artist_id )
     REFERENCES  cs421g22.artists( artist_id ),
@@ -95,8 +95,8 @@ CREATE TABLE cs421g22.song_by_artist (
 );
 
 CREATE TABLE cs421g22.song_in_album (
-    song_id     BIGINT NOT NULL,
-    album_id    BIGINT NOT NULL,
+    song_id     INT NOT NULL,
+    album_id    INT NOT NULL,
     CONSTRAINT  idx_song_in_album_1 UNIQUE ( song_id, album_id ) ,
     CONSTRAINT  fk_song_in_album FOREIGN KEY ( song_id )
     REFERENCES  cs421g22.songs( song_id ),
@@ -105,8 +105,8 @@ CREATE TABLE cs421g22.song_in_album (
 );
 
 CREATE TABLE cs421g22.song_in_playlist (
-    song_id     BIGINT NOT NULL,
-    playlist_id BIGINT NOT NULL,
+    song_id     INT NOT NULL,
+    playlist_id INT NOT NULL,
     CONSTRAINT  idx_song_in_playlist UNIQUE ( song_id, playlist_id ) ,
     CONSTRAINT  fk_song_in_collection FOREIGN KEY ( playlist_id )
     REFERENCES  cs421g22.playlists( playlist_id ),
@@ -115,8 +115,8 @@ CREATE TABLE cs421g22.song_in_playlist (
 );
 
 CREATE TABLE cs421g22.user_likes_playlist (
-    user_id     BIGINT NOT NULL,
-    playlist_id BIGINT NOT NULL,
+    user_id     INT NOT NULL,
+    playlist_id INT NOT NULL,
     CONSTRAINT  idx_user_likes_playlist_1 UNIQUE ( playlist_id, user_id ) ,
     CONSTRAINT  fk_user_likes_playlist FOREIGN KEY ( playlist_id )
     REFERENCES  cs421g22.playlists( playlist_id )
@@ -127,8 +127,8 @@ CREATE TABLE cs421g22.user_likes_playlist (
 );
 
 CREATE TABLE cs421g22.user_has_playlist (
-    user_id     BIGINT NOT NULL,
-    playlist_id BIGINT NOT NULL,
+    user_id     INT NOT NULL,
+    playlist_id INT NOT NULL,
     since       timestamp DEFAULT current_timestamp,
     CONSTRAINT  idx_user_has_playlist_1 UNIQUE ( playlist_id, user_id ) ,
     CONSTRAINT  fk_user_has_playlist FOREIGN KEY ( playlist_id )
